3rd programs contains the gui
integrate those algorithms into a gui. the gui must contain a button which allows the user to choos out fasta file. a fasta file is the main biological file format which can store dna, rna for protein sequences. the first line in the fasta file is the information line which contains an identifier, the species the sequence belongs to and other info regarding the line. starting from the 2nd line is the biological sequence which is split into 80 character lines